backend/pyproject.toml
[tool.poetry]
name = "mhnu-backend"
version = "0.1.0"
description = ""
authors = ["Adrian González <adrian.gonzalez@unillanos.edu.co>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = ">=3.2,<4.0"
psycopg2-binary = "^2.9.3"
django-admin-interface = "^0.18.5"
django-import-export = "^2.7.1"
djangorestframework = "^3.13.1"
drf-yasg = "^1.20.0"
django-rosetta = "^0.9.8"
Pillow = "^9.0.0"
celery = "^5.2.3"
redis = "^4.1.2"
django-cleanup = "^6.0.0"
django-crum = "^0.7.9"
django-object-actions = "^3.1.0"
weasyprint = "^54.1"
django-qr-code = "^3.0.0"
django-autocompletefilter = "^0.0.8"
django-tabbed-admin = "^1.0.4"
django-filter = "^21.1"
django-cors-headers = "^3.11.0"
djangorestframework-camel-case = "^1.3.0"
django-json-widget = "^1.1.1"
gunicorn = "^20.1.0"
drf-recaptcha = "^2.0.7"
django-ckeditor = "^6.2.0"
django-autoslug = "^1.9.8"

[tool.poetry.dev-dependencies]
django-debug-toolbar = "^3.2.4"
watchdog = "^2.1.6"
PyYAML = "^6.0"
argh = "^0.26.2"
django-extensions = "^3.1.5"
pytest = "^6.2.5"
pytest-django = "^4.5.2"
pytest-cov = "^3.0.0"
pytest-xdist = "^2.5.0"
pytest-sugar = "^0.9.4"
black = "^22.1.0"
flake8 = "^4.0.1"
isort = "^5.10.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
skip = ["migrations"]
